how does this work?

+++ web/prove/06/index.php
@@ -47,7 +47,7 @@
             $stmt = $db->prepare('INSERT INTO service_service
                                  VALUES (:id, :vin, :email, :notes, :tech)');
             $stmt->bindParam(':id', htmlspecialchars($ns_id));
-            $stmt->bindParam(':vin', (int)htmlspecialchars($ns_vin));
+            $stmt->bindParam(':vin', htmlspecialchars($ns_vin));
             $stmt->bindParam(':email', htmlspecialchars($ns_email));
             $stmt->bindParam(':notes', htmlspecialchars($ns_notes));
             $stmt->bindParam(':tech', htmlspecialchars($ns_tech));
